requirements.txt
streamlit
langchain-core
pydantic
pyyaml
requests
python-dotenv
httpx
google-generativeai